Make the code work
